[ FeaturePolicy ] 'sandbox' vs 'allow' attributes
This CL ensures that sandbox features set through 'allow' attribute take
precedence over the 'sandbox' attribute itsef. In line with this, any sandbox
flag which is converted to a container policy will be removed from the set of
flags for the corresponding frame owner. This ensures that the container policy
will not be overwritten by FrameOwner sandbox at the time of receiving the
feature policy headers.

The CL also packs a WPT which verifies the precedence of 'allow' over 'sandbox'.

Bug: 812381, 926293
Change-Id: Ic1d31dfb17ce4a81b5ead23b789119c04cfacd8d