Cleanup of backdrop filter code, plus fixes for Mac
Prior to this CL, on a high-DPI Mac, the clip rect would not be placed
in the correct location. It was inverted, and not scaled by the
device_scale_factor.

In addition, this CL changes the naming from "BackgroundFilter" to
"BackdropFilter" within the renderer code, and cleans up the unused
quad parameter on the ShouldApplyBackgroundFilters() function.

Change-Id: I375492d1f6b8d7a1acdc41401f2c00cfeb32d4ec