[css-grid] Treat minimum size percentages as 0 for the minimum contribution
If the computed preferred size of a grid item behaves as auto, its
minimum contribution is the outer size that would result from assuming
the item’s used minimum size as its preferred size.

Before this change, Blink used the computed minimum size instead of the
used one. This is noticeable because cyclic percentages are sometimes
resolved differently as a minimum size than as a preferred size.

This caused problems especially when dynamically changing the minimum
size, then percentages were resolved with respect to the old size of the
grid area.

BUG=942811
TEST=external/wpt/css/css-grid/grid-items/grid-item-dynamic-min-contribution.html
TEST=external/wpt/css/css-grid/grid-items/grid-item-percentage-sizes-003.html

Change-Id: Ic39faad51fef17ef4d2994b60387ea67bc46bef7